+++ LittleAgenda/Controllers/TelephonesController.cs
@@ -74,10 +74,11 @@ namespace LittleAgenda.Controllers
 				db.Telefones.Add(telephone);
 
 				await db.SaveChangesAsync();
-                return RedirectToAction("Index");
-            }
 
-            ViewBag.ContactId = new SelectList(db.Contatos, "ContactId", "Name", telephone.ContactId);
+				return new HttpStatusCodeResult(HttpStatusCode.OK);
+			}
+
+            ViewBag.ContactId = telephone.ContactId;
             return View(telephone);
         }
 
@@ -93,7 +94,7 @@ namespace LittleAgenda.Controllers
             {
                 return HttpNotFound();
             }
-            ViewBag.ContactId = new SelectList(db.Contatos, "ContactId", "Name", telephone.ContactId);
+            ViewBag.ContactId = telephone.ContactId;
             return View(telephone);
         }
 
@@ -112,8 +113,9 @@ namespace LittleAgenda.Controllers
 				db.Entry(telephone).State = EntityState.Modified;
 
 				await db.SaveChangesAsync();
-                return RedirectToAction("Index");
-            }
+
+				return new HttpStatusCodeResult(HttpStatusCode.OK);
+			}
             ViewBag.ContactId = new SelectList(db.Contatos, "ContactId", "Name", telephone.ContactId);
             return View(telephone);
         }
@@ -138,15 +140,24 @@ namespace LittleAgenda.Controllers
         [ValidateAntiForgeryToken]
         public async Task<ActionResult> DeleteConfirmed(string id)
         {
-            Telephone telephone = await db.Telefones.FindAsync(id);
+			if (id == null)
+			{
+				return new HttpStatusCodeResult(HttpStatusCode.BadRequest);
+			}
+			Telephone telephone = await db.Telefones.FindAsync(id);
+			if (telephone == null)
+			{
+				return HttpNotFound();
+			}
 
 			var contact = db.Contatos.Find(telephone.ContactId);
 			telephone.Contact = contact;
 
 			db.Telefones.Remove(telephone);
             await db.SaveChangesAsync();
-            return RedirectToAction("Index");
-        }
+
+			return new HttpStatusCodeResult(HttpStatusCode.OK);
+		}
 
         protected override void Dispose(bool disposing)
         {
